fix: rename IsTaxID file to dash-case
