+++ src/decorator/decorators.ts
@@ -114,7 +114,7 @@ export * from './string/IsSemVer';
 export * from './string/IsStrongPassword';
 export * from './string/IsTimeZone';
 export * from './string/IsBase58';
-export * from './string/IsTaxId';
+export * from './string/is-tax-id';
 
 // -------------------------------------------------------------------------
 // Type checkers
